Implementations of the custom layers i.e the Correlation layer, Resample2D layer and the ChannelNorm layer in native Pytorch and C++ to support inference on CPU.
The main bottleneck is the Correlation Layer on which the FlowNetC architecture relies.
This PR provides 2 implementations of the Correlation layer.

-PyTorch native implementation. This requires no extra setup
-Optimized C++ implementation for inference on CPU.

Also provided are Pytorch native implementations for Resample2D and Channelnorm.
Since the Pytorch implementation is quite efficient (compeletely vectorized) with no
python for loops, C++ implementation is not needed. These layers also work by default
on the GPU dependening on if the input tensors are on gpu and are slightly slower than
the provided cuda implementation.

See comments at the top of models.py and networks/FlowNetC.py for more details and
how to switch to CPU mode.

Backward passes are not yet implemented but will be added in the future.

run_a_pair.py is replaced with a generic script called test.py to simply test functionality.
run_a_pair.py had hardcoded paths. Also 2 frames from sintel added in test_images dir so that
functionality and setup can be swiftly checked.

Resolves: #190